# Fix performance-type-promotion-in-math-fn clang-tidy warning